Add arduino wifi 2 board

+++ inst/+arduinoio/lib/OctaveI2CLibrary.cpp
@@ -73,7 +73,7 @@ void OctaveI2CLibrary::commandHandler(uint8_t cmdID, uint8_t* data, uint8_t data
 	  sendInvalidNumArgsMsg();
         }
         else {
-          Wire.requestFrom(data[0], data[1]);
+          Wire.requestFrom(data[0], (size_t)data[1]);
           byte c = 0;
           byte l = data[1];
 	  if(l > 5) sendWaitMsg();
@@ -111,7 +111,7 @@ void OctaveI2CLibrary::commandHandler(uint8_t cmdID, uint8_t* data, uint8_t data
           }
           Wire.endTransmission(false);
           byte l = data[2+data[1]];
-          Wire.requestFrom(data[0], l);
+          Wire.requestFrom(data[0], (size_t)l);
 	  if(l > 5) sendWaitMsg();
           datasz = 2;
           for(c=0;c<=l;c++)
